requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
alembic
openai
passlib[bcrypt]
python-jose[cryptography]
